requirements.txt
autopep8==1.3.3
byteplay3==3.5.0
neo-boa==0.2.1
pycodestyle==2.3.1